+++ ca-api/src/main/java/org/xipki/ca/api/profile/ExtensionSpec.java
@@ -311,7 +311,6 @@ public abstract class ExtensionSpec {
 
     private static final Set<ASN1ObjectIdentifier> REQUIRED_EXTENSIONS =
         Collections.unmodifiableSet(new HashSet<>(Arrays.asList(
-            Extension.authorityInfoAccess,
             Extension.subjectKeyIdentifier
             )));